+++ src/Differences.DataAccess/Migrations/DifferencesDbContextModelSnapshot.cs
@@ -23,18 +23,18 @@ namespace Differences.DataAccess.Migrations
 
             modelBuilder.Entity("Differences.Interaction.Models.Article", b =>
                 {
-                    b.Property<long>("Id")
+                    b.Property<int>("Id")
                         .IsConcurrencyToken()
                         .ValueGeneratedOnAdd();
 
-                    b.Property<long>("AuthorId");
+                    b.Property<int>("AuthorId");
 
                     b.Property<string>("Content")
                         .IsRequired();
 
                     b.Property<DateTime>("CreateTime");
 
-                    b.Property<long>("CreatedBy");
+                    b.Property<int>("CreatedBy");
 
                     b.Property<DateTime?>("LastUpdateTime");
 
@@ -51,18 +51,18 @@ namespace Differences.DataAccess.Migrations
 
             modelBuilder.Entity("Differences.Interaction.Models.ArticleUpdateHistory", b =>
                 {
-                    b.Property<long>("Id")
+                    b.Property<int>("Id")
                         .IsConcurrencyToken()
                         .ValueGeneratedOnAdd();
 
-                    b.Property<long>("ArticleId");
+                    b.Property<int>("ArticleId");
 
                     b.Property<string>("Content")
                         .IsRequired();
 
                     b.Property<DateTime>("CreateTime");
 
-                    b.Property<long>("CreatedBy");
+                    b.Property<int>("CreatedBy");
 
                     b.Property<DateTime?>("LastUpdateTime");
 
@@ -77,11 +77,11 @@ namespace Differences.DataAccess.Migrations
 
             modelBuilder.Entity("Differences.Interaction.Models.Comment", b =>
                 {
-                    b.Property<long>("Id")
+                    b.Property<int>("Id")
                         .IsConcurrencyToken()
                         .ValueGeneratedOnAdd();
 
-                    b.Property<long>("ArticleId");
+                    b.Property<int>("ArticleId");
 
                     b.Property<string>("Content")
                         .IsRequired()
@@ -89,13 +89,13 @@ namespace Differences.DataAccess.Migrations
 
                     b.Property<DateTime>("CreateTime");
 
-                    b.Property<long>("CreatedBy");
+                    b.Property<int>("CreatedBy");
 
                     b.Property<DateTime?>("LastUpdateTime");
 
-                    b.Property<long>("OwnerId");
+                    b.Property<int>("OwnerId");
 
-                    b.Property<long?>("ParentCommentId");
+                    b.Property<int?>("ParentCommentId");
 
                     b.HasKey("Id");
 
@@ -108,7 +108,7 @@ namespace Differences.DataAccess.Migrations
 
             modelBuilder.Entity("Differences.Interaction.Models.Question", b =>
                 {
-                    b.Property<long>("Id")
+                    b.Property<int>("Id")
                         .IsConcurrencyToken()
                         .ValueGeneratedOnAdd();
 
@@ -118,11 +118,11 @@ namespace Differences.DataAccess.Migrations
 
                     b.Property<DateTime>("CreateTime");
 
-                    b.Property<long>("CreatedBy");
+                    b.Property<int>("CreatedBy");
 
                     b.Property<DateTime?>("LastUpdateTime");
 
-                    b.Property<long>("OwnerId");
+                    b.Property<int>("OwnerId");
 
                     b.Property<string>("Title")
                         .IsRequired()
@@ -137,7 +137,7 @@ namespace Differences.DataAccess.Migrations
 
             modelBuilder.Entity("Differences.Interaction.Models.Reply", b =>
                 {
-                    b.Property<long>("Id")
+                    b.Property<int>("Id")
                         .IsConcurrencyToken()
                         .ValueGeneratedOnAdd();
 
@@ -147,15 +147,15 @@ namespace Differences.DataAccess.Migrations
 
                     b.Property<DateTime>("CreateTime");
 
-                    b.Property<long>("CreatedBy");
+                    b.Property<int>("CreatedBy");
 
                     b.Property<DateTime?>("LastUpdateTime");
 
-                    b.Property<long>("OwnerId");
+                    b.Property<int>("OwnerId");
 
-                    b.Property<long?>("ParentReplyId");
+                    b.Property<int?>("ParentReplyId");
 
-                    b.Property<long>("QuestionId");
+                    b.Property<int>("QuestionId");
 
                     b.HasKey("Id");
 
@@ -168,7 +168,7 @@ namespace Differences.DataAccess.Migrations
 
             modelBuilder.Entity("Differences.Interaction.Models.User", b =>
                 {
-                    b.Property<long>("Id")
+                    b.Property<int>("Id")
                         .IsConcurrencyToken()
                         .ValueGeneratedOnAdd();
 
@@ -177,7 +177,7 @@ namespace Differences.DataAccess.Migrations
 
                     b.Property<DateTime>("CreateTime");
 
-                    b.Property<long>("CreatedBy");
+                    b.Property<int>("CreatedBy");
 
                     b.Property<string>("DisplayName")
                         .IsRequired()
